+++ ecr_chess/src/pieces/knight.rs
@@ -1,9 +1,9 @@
-use crate::pieces::{PieceColor, PieceType};
-
-use super::Piece;
 use crate::board::Board;
 use crate::coordinate::Coordinate;
-use crate::pieces::move_gen::{knight_moves, BasicMove};
+use crate::pieces::{PieceColor, PieceType};
+use crate::pieces::move_gen::{BasicMove, knight_moves};
+
+use super::Piece;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Knight {}
@@ -22,7 +22,7 @@ impl Piece for Knight {
         knight_moves(piece_coordinate, board, piece_color)
     }
 
-    fn get_value(&self) -> usize {
+    fn get_value(&self) -> u8 {
         30
     }
 }
