+++ ecr_chess/src/pieces/pawn.rs
@@ -1,9 +1,9 @@
-use crate::pieces::{PieceColor, PieceType};
-
-use super::Piece;
 use crate::board::Board;
 use crate::coordinate::Coordinate;
-use crate::pieces::move_gen::{pawn_moves, BasicMove};
+use crate::pieces::{PieceColor, PieceType};
+use crate::pieces::move_gen::{BasicMove, pawn_moves};
+
+use super::Piece;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Pawn {}
@@ -22,7 +22,7 @@ impl Piece for Pawn {
         pawn_moves(piece_coordinate, board, piece_color, has_moved)
     }
 
-    fn get_value(&self) -> usize {
+    fn get_value(&self) -> u8 {
         10
     }
 }
