+++ ecr_chess/src/pieces/rook.rs
@@ -1,9 +1,9 @@
-use crate::pieces::{PieceColor, PieceType};
-
-use super::Piece;
 use crate::board::Board;
 use crate::coordinate::Coordinate;
-use crate::pieces::move_gen::{linear_moves, BasicMove};
+use crate::pieces::{PieceColor, PieceType};
+use crate::pieces::move_gen::{BasicMove, linear_moves};
+
+use super::Piece;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Rook {}
@@ -22,7 +22,7 @@ impl Piece for Rook {
         linear_moves(piece_coordinate, board, piece_color)
     }
 
-    fn get_value(&self) -> usize {
+    fn get_value(&self) -> u8 {
         50
     }
 }
